Remove wiiupc, libogc wiiuse supports WiiU Pro controller - Part 1
Wii U Pro controller is natively supported since libogc 1.8.15-1 [Tantric]

+++ source/filebrowser.cpp
@@ -24,6 +24,7 @@
 #include <di/di.h>
 #endif
 
+#include "snes9x/port.h"
 #include "snes9xgx.h"
 #include "filebrowser.h"
 #include "menu.h"
@@ -449,7 +450,7 @@ int BrowserLoadSz()
 
 int WiiFileLoader()
 {
-	int size;
+	size_t size;
 	char filepath[1024];
 
 	memset(Memory.NSRTHeader, 0, sizeof(Memory.NSRTHeader));
